+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * None.
 
 ### Fixed
+* [RealmApp] Failing to refresh the access token due to a 401/403 error will now correctly emit an error with `ErrorCode.BAD_AUTHENTICATION` rather than `ErrorCode.PERMISSION_DENIED`. (Realm Core [#4881](https://github.com/realm/realm-core/issues/4881), since 10.6.1)
+* [RealmApp] If an object with a null primary key was deleted by another sync client, the exception `KeyNotFound: No such object` could be triggered. ([Realm Core #4885](https://github.com/realm/realm-core/issues/4885), since 10.0.0)
 * Exceptions inside change listeners running on background looper threads would crash the Looper with a native `JNI DETECTED ERROR IN APPLICATION: JNI NewLocalRef called with pending exception` instead of the original Java exception. This could also happen when canceling a corutine using a background looper as a Dispatcher.
 
 ### Compatibility
@@ -12,7 +14,7 @@
 * Realm Studio 11.0.0-alpha.0 or above is required to open Realms created by this version.
 
 ### Internal
-* Updated to Realm Core 11.3.1, commit: 15219f3b4b31b565c6a312cfa52f98821cafe798.
+* Updated to Realm Core 11.4.0, commit: 9b2f67c24581503486d8e1d2066fd7e8c5fd1491.
 
 
 ## 10.8.0 (2021-08-27)