+++ CHANGELOG.md
@@ -13,7 +13,8 @@
 
 ### Internal
 * Upgrade to JDK 11.
-* Updated to Gradle 7.1.1.
+* Updated to Gradle 7.2.
+* Updated to Android Gradle Plugin 4.1.3.
 * Disable analytics for any value of the `REALM_DISABLE_ANALYTICS` environment variable, not just `true`.
 * Disable analytics whenever the `CI` environment variable is set.